requirements.txt
streamlit
python-dotenv
msal